Add files via upload
In this program, the password is categorized into three categories: invalid, weak, and strong. The password becomes invalid if the length of the password is less than 8 characters. The password falls under the weak category if it does not contain at least one uppercase character, one lowercase character, one digit, and a special character. And the password comes under the strong category if it contains all these characters.